requirements.txt
argcomplete==1.10.3
beautifulsoup4==4.8.2
chardet==3.0.4
compressed_rtf==1.0.6
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
IMAPClient==2.1.0
lxml==5.1.0
olefile==0.47
pdfminer.six==20191110
pillow==10.2.0
pycryptodome==3.19.1
PyPDF2==3.0.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-pptx==0.6.23
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
textract==1.6.5
tzdata==2023.4
tzlocal==5.2
xlrd==1.2.0
XlsxWriter==3.1.9